new region file
